[Enhancement] Consolidate signup charts into User Activity page
Move the signup activity calendar and cumulative user growth charts from
the standalone Charts page into a new Signups section on the User
Activity page, extracted as a SignupCharts component. Remove the Charts
route and sidebar entry; the old /admin/dashboard/charts path redirects
to /admin/dashboard/activity.
